requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
asgiref==3.10.0
    # via django
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
django==5.2.7
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-rest-framework==0.1.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   django-rest-framework
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
psycopg2-binary==2.9.11
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dotenv==1.1.1
    # via -r requirements.in
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via referencing
uritemplate==4.2.0
    # via drf-spectacular
